reordering to avoid messing up my offset.
I'd done putting the <q> after the opening problem tag first, and hadn't thought about how that meant when I used my saved start index thing it would end up in a slightly different place. This turned out to be a problem for multiparts (I ran into one) with everything in the question tags, bc it meant the second to last close was followed directly by the last open so there wasn't room for error. 
Fixed with reordering.